Move normalization from config to source
Instead of setting normalization in the config parameter,
pass normalization as a parameter to Source initialization.
