+++ scarlet/source.py
@@ -3,7 +3,7 @@ import numpy as np
 
 import proxmin
 from . import constraint as sc
-from .config import Config, Normalization
+from .config import Config
 from .component import Component, ComponentTree
 from .operator import prox_sed_on
 
@@ -138,7 +138,8 @@ class PointSource(Source):
     symmetry and monotonicity.
     """
     def __init__(self, center, img, shape=None, constraints=None, psf=None, config=None,
-                 fix_sed=False, fix_morph=False, fix_frame=False, shift_center=0.1, tiny=1e-10):
+                 fix_sed=False, fix_morph=False, fix_frame=False, shift_center=0.1, tiny=1e-10,
+                 normalization=sc.Normalization.A):
         """Initialize
 
         This implementation initializes the sed from the pixel in
@@ -157,6 +158,8 @@ class PointSource(Source):
             Shape of the initial morphology.
             If `shape` is `None` then the smallest shape specified by `config.source_sizes`
             is used.
+        normalization: `Normalization`, default = `Normalization.A`
+            The normalization method used to break the AS degeneracy.
         """
         if config is None:
             config = Config()
@@ -164,12 +167,14 @@ class PointSource(Source):
             shape = (config.source_sizes[0],) * 2
         sed, morph = self._make_initial(center, img, shape, psf, config, tiny)
 
+        self.normalization = normalization
         if constraints is None:
-            constraints = (sc.SimpleConstraint(config.normalization),
+            constraints = (sc.SimpleConstraint(normalization),
                            sc.DirectMonotonicityConstraint(use_nearest=False),
                            sc.DirectSymmetryConstraint())
 
-        component = Component(sed, morph, center=center, constraints=constraints, psf=psf, fix_sed=fix_sed, fix_morph=fix_morph, fix_frame=fix_frame, shift_center=shift_center)
+        component = Component(sed, morph, center=center, constraints=constraints, psf=psf, fix_sed=fix_sed,
+                              fix_morph=fix_morph, fix_frame=fix_frame, shift_center=shift_center)
         super(PointSource, self).__init__(component)
 
     def _make_initial(self, center, img, shape, psf, config, tiny=1e-10):
@@ -186,7 +191,7 @@ class PointSource(Source):
         """
         B, Ny, Nx = img.shape
         _y, _x = center_int = np.round(center).astype('int')
-        if config.normalization != Normalization.A:
+        if self.normalization != sc.Normalization.A:
             sed = img[:, _y, _x]
             if psf is not None:
                 # Increase the magnitude of the SED's to account
@@ -227,7 +232,8 @@ class ExtendedSource(Source):
     but other `constraints` can be used.
     """
     def __init__(self, center, img, bg_rms, constraints=None, psf=None, symmetric=True, monotonic=True,
-                 thresh=1., config=None, fix_sed=False, fix_morph=False, fix_frame=False, shift_center=0.2):
+                 thresh=1., config=None, fix_sed=False, fix_morph=False, fix_frame=False, shift_center=0.2,
+                 normalization=sc.Normalization.A):
         """Initialize
 
         See :class:`~scarlet.source.Source` for parameter descriptions not listed below.
@@ -247,6 +253,8 @@ class ExtendedSource(Source):
         thresh: float
             Multiple of the RMS used to set the minimum non-zero flux.
             Use `thresh=1` to just use `bg_rms` to set the flux floor.
+        normalization: `Normalization`, default = `Normalization.A`
+            The normalization method used to break the AS degeneracy.
         """
         # Use a default configuration if config is not specified
         if config is None:
@@ -255,8 +263,9 @@ class ExtendedSource(Source):
         sed, morph = self._make_initial(center, img, bg_rms, thresh=thresh, symmetric=symmetric,
                                         monotonic=monotonic, config=config)
 
+        self.normalization = normalization
         if constraints is None:
-            constraints = (sc.SimpleConstraint(config.normalization),
+            constraints = (sc.SimpleConstraint(normalization),
                            sc.DirectMonotonicityConstraint(use_nearest=False),
                            sc.DirectSymmetryConstraint())
 
@@ -305,10 +314,10 @@ class ExtendedSource(Source):
         # For the different S matrix normalizations,
         # we make sure that our inital values are in agreement
         # with the correct normalization
-        if config.normalization != Normalization.A:
-            if config.normalization == Normalization.S:
+        if self.normalization != sc.Normalization.A:
+            if self.normalization == sc.Normalization.S:
                 norm = np.sum(morph)
-            elif config.normalization == Normalization.Smax:
+            elif self.normalization == sc.Normalization.Smax:
                 norm = np.max(morph)
             sed *= norm
             morph /= norm
@@ -391,8 +400,9 @@ class MultiComponentSource(ExtendedSource):
     The SED for all components is calculated as the best fit of the multi-component
     morphology to the multi-band image in the region of the source.
     """
-    def __init__(self, center, img, bg_rms, size_percentiles=[50], constraints=None, psf=None, symmetric=True, monotonic=True,
-                 thresh=1., config=None, fix_sed=False, fix_morph=False, fix_frame=False, shift_center=0.2):
+    def __init__(self, center, img, bg_rms, size_percentiles=[50], constraints=None, psf=None,
+                 symmetric=True, monotonic=True, thresh=1., config=None, fix_sed=False, fix_morph=False,
+                 fix_frame=False, shift_center=0.2, normalization=sc.Normalization.A):
         """Initialize multi-component source, where the inner components begin
         at the given size_percentiles.
         See `~scarlet.source.ExtendedSource` for details.
@@ -401,13 +411,17 @@ class MultiComponentSource(ExtendedSource):
         if config is None:
             config = Config()
 
+        self.normalization = normalization
         if constraints is None:
-            constraints = (sc.SimpleConstraint(),
+            constraints = (sc.SimpleConstraint(normalization),
                            sc.DirectMonotonicityConstraint(use_nearest=False),
                            sc.DirectSymmetryConstraint())
 
         # start from ExtendedSource for single-component morphology and sed
-        super(MultiComponentSource, self).__init__(center, img, bg_rms, constraints=constraints, psf=psf, symmetric=symmetric, monotonic=monotonic, thresh=thresh, config=config, fix_sed=fix_sed, fix_morph=fix_morph, fix_frame=fix_frame, shift_center=shift_center)
+        super(MultiComponentSource, self).__init__(center, img, bg_rms, constraints=constraints, psf=psf,
+                                                   symmetric=symmetric, monotonic=monotonic, thresh=thresh,
+                                                   config=config, fix_sed=fix_sed, fix_morph=fix_morph,
+                                                   fix_frame=fix_frame, shift_center=shift_center)
 
         # create a list of components from base morph by layering them on top of
         # each other so that they sum up to morph
